Prepares for distribution with Replicated
TL;DR
-----

Adds the Replicated SDK to enable distribution with Replicated

Details
-------

Updates the chart to include a dependency on the Replicated SDK. This
dependency is all that's need to support integration with the Replicated
Platform. It enables Replicated to inject licensing information and to
collect telemetry about running instances.