Proxies images through the Replicated proxy (#4)
TL;DR
-----

Uses the Replicated proxy registry

Details
-------

Proxies images through the Replicated proxy registry and adds the required
image pull secret. Right now the secret is named `replicated-registry` since it
the Bitnami common library forces us to specify a static name in
`global.imagePullSeecrets`. This was the quickest way to get it going and
remain idiomatic to the way Bitnami does thingss, but it makes the secret
inconsistent with the usually naming Bitnami naming conventions.