Continuously deliver to the Replicated Platform with GitHub or GitLab (#5)
TL;DR
-----

Adds CD pipelines for GitHub and GitLab

Details
-------

Implements initial CD capabilities using a GitHub Actions workflow and GitLab
CI/CD. This is a baseline to add more pipelines as we encounter customers who
use different CI/CD platforms.

The GitLab pipeline is a bit more complicated than it needs to be because the
current implementation of the `replicated` CLI makes it difficult to script out
scenarios where you split releasing from promoting. I opted to do that for
both the visual impact and because it is a more aligned with the idea of
testing a release _before_ releasing it, which we'll do in the next iteration.